langgraph-examples/pyproject.toml
[project]
name = "langgraph-examples"
version = "0.1.0"
description = "LangGraph implementations of The Agents Playbook examples"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "langchain-anthropic>=1.4.2",
    "langchain-core>=1.3.2",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.10",
    "numpy>=2.4.4",
    "pydantic-settings>=2.14.0",
]

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
